This pull requests completes fix of PR #12474 regarding an integer overflow in processTimeEvents of ae.c
Although the major effect of maximizing the CPU usage was handled, the integer overflow could still happen for calculated next periods > INT_MAX resulting in wrong calculation of the next period of the time event

To address this, prototype was changed to return a long long in order to prevent the above overflow

+++ src/ae.h
@@ -43,7 +43,7 @@ struct aeEventLoop;
 
 /* Types and data structures */
 typedef void aeFileProc(struct aeEventLoop *eventLoop, int fd, void *clientData, int mask);
-typedef int aeTimeProc(struct aeEventLoop *eventLoop, long long id, void *clientData);
+typedef long long aeTimeProc(struct aeEventLoop *eventLoop, long long id, void *clientData);
 typedef void aeEventFinalizerProc(struct aeEventLoop *eventLoop, void *clientData);
 typedef void aeBeforeSleepProc(struct aeEventLoop *eventLoop);
